+++ .env.example
@@ -1,2 +1,3 @@
+NEXT_PUBLIC_INSTANT_APP_ID=
 INSTANT_ADMIN_TOKEN=
 SECRET=
