+++ src/lib/types.ts
@@ -1,0 +1,9 @@
+export type Schema = {
+  body: { markdown: string };
+  posts: {
+    title: string;
+    createdAt: number;
+    updatedAt: number;
+    number: number;
+  }
+}
